# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Abaya Production Tracking and Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /master/categories:
    get:
      operationId: listCategories
      tags: [master]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [master]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /master/categories/{id}:
    put:
      operationId: updateCategory
      tags: [master]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [master]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /master/sizes:
    get:
      operationId: listSizes
      tags: [master]
      summary: List all sizes
      responses:
        "200":
          description: List of sizes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Size"
    post:
      operationId: createSize
      tags: [master]
      summary: Create a size
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSizeBody"
      responses:
        "201":
          description: Created size
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Size"
  /master/colors:
    get:
      operationId: listColors
      tags: [master]
      summary: List all colors
      responses:
        "200":
          description: List of colors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Color"
    post:
      operationId: createColor
      tags: [master]
      summary: Create a color
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateColorBody"
      responses:
        "201":
          description: Created color
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Color"
  /master/colors/{id}:
    put:
      operationId: updateColor
      tags: [master]
      summary: Update a color
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateColorBody"
      responses:
        "200":
          description: Updated color
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Color"
  /master/materials:
    get:
      operationId: listMaterials
      tags: [master]
      summary: List all materials (accessories)
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [master]
      summary: Create a material (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaterialBody"
      responses:
        "201":
          description: Created material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /master/materials/{id}:
    put:
      operationId: updateMaterial
      tags: [master]
      summary: Update a material (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMaterialBody"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /master/fabrics:
    get:
      operationId: listFabrics
      tags: [master]
      summary: List all fabrics
      responses:
        "200":
          description: List of fabrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fabric"
    post:
      operationId: createFabric
      tags: [master]
      summary: Create a fabric type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFabricBody"
      responses:
        "201":
          description: Created fabric
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fabric"
  /master/fabrics/{id}:
    put:
      operationId: updateFabric
      tags: [master]
      summary: Update a fabric type (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFabricBody"
      responses:
        "200":
          description: Updated fabric
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fabric"
  /master/products:
    get:
      operationId: listProducts
      tags: [master]
      summary: List all products/designs
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [master]
      summary: Create a product/design
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /master/stitchers:
    get:
      operationId: listStitchers
      tags: [master]
      summary: List all stitchers
      responses:
        "200":
          description: List of stitchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stitcher"
    post:
      operationId: createStitcher
      tags: [master]
      summary: Create a stitcher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStitcherBody"
      responses:
        "201":
          description: Created stitcher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stitcher"
  /master/stitchers/{id}:
    put:
      operationId: updateStitcher
      tags: [master]
      summary: Update a stitcher (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStitcherBody"
      responses:
        "200":
          description: Updated stitcher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stitcher"
  /master/teams:
    get:
      operationId: listTeams
      tags: [master]
      summary: List all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [master]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamBody"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /master/teams/{id}:
    put:
      operationId: updateTeam
      tags: [master]
      summary: Update a team (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamBody"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /master/users:
    get:
      operationId: listUsers
      tags: [master]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [master]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
  /master/users/{id}:
    put:
      operationId: updateUser
      tags: [master]
      summary: Update user role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"

  # FABRIC ROLLS
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required: [id, name]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
    Size:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
      required: [id, name]
    CreateSizeBody:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer
      required: [name]
    Color:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
      required: [id, name]
    CreateColorBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
          description: Short unique color code, e.g. BLK, NVY, BRN
      required: [name, code]
    UpdateColorBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
          description: Short unique color code, e.g. BLK, NVY, BRN
      required: [name, code]
    Material:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, code, name, isActive]
    CreateMaterialBody:
      type: object
      properties:
        code:
          type: string
          description: Short unique material code, e.g. LC01, DR02
        name:
          type: string
        description:
          type: string
      required: [code, name]
    UpdateMaterialBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
      required: [code, name]
    Fabric:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
          nullable: true
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        isActive:
          type: boolean
      required: [id, name]
    CreateFabricBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        unit:
          type: string
      required: [name]
    UpdateFabricBody:
      type: object
      properties:
        code:
          type: string
          nullable: true
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        isActive:
          type: boolean
    Product:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        description:
          type: string
        isActive:
          type: boolean
      required: [id, name]
    CreateProductBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        categoryId:
          type: integer
        description:
          type: string
      required: [name]
    Stitcher:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        phone:
          type: string
        teamId:
          type: integer
        teamName:
          type: string
        isActive:
          type: boolean
        totalIssued:
          type: integer
        totalReceived:
          type: integer
        pendingQuantity:
          type: integer
      required: [id, name]
    CreateStitcherBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        phone:
          type: string
        teamId:
          type: integer
      required: [name]
    UpdateStitcherBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        phone:
          type: string
        teamId:
          type: integer
        isActive:
          type: boolean
      required: [name]
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        supervisorName:
          type: string
        isActive:
          type: boolean
      required: [id, name]
    CreateTeamBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        supervisorName:
          type: string
      required: [name]
    UpdateTeamBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        supervisorName:
          type: string
        isActive:
          type: boolean
      required: [name]
    AppUser:
      type: object
      properties:
        id:
          type: integer
        replitUserId:
          type: string
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, username, role]
    CreateUserBody:
      type: object
      properties:
        replitUserId:
          type: string
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
      required: [username, role]
    UpdateUserBody:
      type: object
      properties:
        role:
          type: string
        isActive:
          type: boolean
      required: [role]

    # FABRIC ROLLS
...