# lib/api-zod/src/generated/types/updateFabricBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Abaya Production Tracking and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateFabricBody {
  code?: string | null;
  name?: string;
  description?: string;
  unit?: string;
  isActive?: boolean;
}